+++ tests/unit/no-top-level-side-effects.test.ts
@@ -371,6 +371,14 @@ const valid: RuleTester.ValidTestCase[] = [
     {
       code: `export const bigInt = BigInt();`,
       options: [options.allowCallBigInt]
+    },
+    {
+      code: `const symbol = Symbol.for("membership");`,
+      options: [{allowedCalls: ['Symbol.for']}]
+    },
+    {
+      code: `const symbol = x.y.z();`,
+      options: [{allowedCalls: ['x.y.z']}]
     }
   ],
 
@@ -1690,6 +1698,32 @@ const invalid: RuleTester.InvalidTestCase[] = [
           column: 1,
           endLine: 1,
           endColumn: 27
+        }
+      ]
+    },
+    {
+      code: `console.log('hello world');`,
+      options: [{allowedCalls: ['console.info']}],
+      errors: [
+        {
+          messageId: '0',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 27
+        }
+      ]
+    },
+    {
+      code: `console[log]('hello world');`,
+      options: [{allowedCalls: ['console.log']}],
+      errors: [
+        {
+          messageId: '0',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 28
         }
       ]
     },